+++ tests/e2e-tests/src/main/kotlin/api_models/Credential.kt
@@ -12,5 +12,7 @@ data class Credential(
     var updatedAt: String = "",
     var validityPeriod: Int = 0,
     var claims: LinkedHashMap<String,String> = LinkedHashMap(),
-    var jwtCredential: String = ""
+    var jwtCredential: String = "",
+    var issuingDID: String = "",
+    var connectionId: String = ""
 )